[window] Move window and surface code to subfolder in wrapper folder
